FacePrivacy/PRO-Face C/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
absl-py=0.15.0=pyhd3eb1b0_0
aiohttp=3.8.1=py38h7f8727e_1
aiosignal=1.2.0=pyhd3eb1b0_0
async-timeout=4.0.2=py38h06a4308_0
attrs=21.4.0=pyhd3eb1b0_0
bcolz=1.2.1=pypi_0
blas=1.0=mkl
blinker=1.4=py38h06a4308_0
blosc=1.21.3=h6a678d5_0
bokeh=2.4.3=py38h06a4308_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py38h0a891b7_1004
brunsli=0.1=h2531618_0
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
c-blosc2=2.10.5=h80c7b02_0
ca-certificates=2023.12.12=h06a4308_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2023.11.17=py38h06a4308_0
cffi=1.15.1=py38h74dc2b5_0
cfitsio=3.470=h5893167_7
charls=2.2.0=h2531618_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.0.4=py38h06a4308_0
cloudpickle=2.0.0=pyhd3eb1b0_0
colorama=0.4.5=pyhd8ed1ab_0
conda=22.9.0=py38h578d9bd_1
conda-package-handling=1.9.0=py38h0a891b7_0
cryptography=38.0.2=py38h2b5fc30_0
cudatoolkit=11.6.0=hecad31d_10
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.11.0=py38h7b6447c_0
dask=2022.2.1=pyhd3eb1b0_0
dask-core=2022.2.1=pyhd3eb1b0_0
dataclasses=0.8=pyh6d0b6a4_7
dbus=1.13.18=hb2f20db_0
distributed=2022.2.1=pyhd3eb1b0_0
einops=0.6.1=pyhd8ed1ab_0
expat=2.4.9=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
fftw=3.3.9=h27cfd23_1
filelock=3.9.0=py38h06a4308_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
frozenlist=1.2.0=py38h7f8727e_0
fsspec=2022.7.1=py38h06a4308_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
grpcio=1.42.0=py38hce63b2e_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
hdf5=1.12.1=h70be1eb_2
heapdict=1.0.1=pyhd3eb1b0_0
huggingface_hub=0.10.1=py38h06a4308_0
icu=58.2=he6710b0_3
idna=3.4=pyhd8ed1ab_0
imagecodecs=2021.8.26=py38hf0132c2_1
imageio=2.19.3=py38h06a4308_0
importlib-metadata=4.11.3=py38h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
jinja2=3.1.2=py38h06a4308_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h7f8727e_0
jxrlib=1.1=h7b6447c_2
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.2=py38h295c915_0
kornia=0.6.12=pyhd8ed1ab_0
krb5=1.19.3=h3790be6_0
lame=3.100=h166bdaf_1003
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libaec=1.0.4=he6710b0_1
libarchive=3.5.2=hb890918_3
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=10.0.1=default_hb85057a_2
libcurl=7.85.0=h7bff187_0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h7f8727e_1
libevent=2.1.12=h8f2d780_0
libffi=3.3=he6710b0_2
libgcc-ng=12.1.0=h8d9b700_16
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libiconv=1.17=h166bdaf_0
libllvm10=10.0.1=hbcb73fb_5
libmamba=0.27.0=h0dd8ff0_0
libmambapy=0.27.0=py38h923e62a_0
libnghttp2=1.47.0=hdcd2b5c_1
libpng=1.6.37=hbc83047_0
libpq=12.9=h16c4e8d_3
libprotobuf=3.20.1=h4ff587b_0
libsolv=0.7.22=h6239696_0
libssh2=1.10.0=h8f2d780_0
libstdcxx-ng=12.1.0=ha89aaad_16
libtiff=4.4.0=hecacb30_0
libuuid=1.0.3=h7f8727e_2
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
libzlib=1.2.12=h166bdaf_3
libzopfli=1.0.3=he6710b0_0
llvm-openmp=14.0.4=he0ac6c6_0
locket=1.0.0=py38h06a4308_0
lz4-c=1.9.4=h6a678d5_0
lzo=2.10=h516909a_1000
mamba=0.27.0=py38h1abaa86_0
markdown=3.3.4=py38h06a4308_0
markupsafe=2.1.3=py38h5eee18b_0
matplotlib=3.5.2=py38h06a4308_0
matplotlib-base=3.5.2=py38hf590b9c_0
menpo=0.11.0=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
msgpack-python=1.0.3=py38hd09550d_0
multidict=6.0.2=py38h5eee18b_0
munkres=1.1.4=py_0
ncurses=6.3=h5eee18b_3
nettle=3.6=he412f7d_0
networkx=2.8.4=py38h06a4308_0
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numexpr=2.8.4=py38he184ba9_0
numpy=1.23.1=py38h6c91a56_0
numpy-base=1.23.1=py38ha15fc14_0
oauthlib=3.2.1=py38h06a4308_0
opencv-python=4.9.0.80=pypi_0
openh264=2.1.1=h780b84a_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1w=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.2.3=py38h51da96c_0
partd=1.2.0=pyhd3eb1b0_1
pcre=8.45=h295c915_0
pillow=9.2.0=py38hace64e9_1
pip=22.2.2=py38h06a4308_0
ply=3.11=py38_0
prettytable=3.5.0=py38h06a4308_0
protobuf=3.20.1=py38h295c915_0
psutil=5.9.0=py38h5eee18b_0
py-cpuinfo=9.0.0=py38h06a4308_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pybind11-abi=4=hd8ed1ab_3
pycosat=0.6.3=py38h0a891b7_1010
pycparser=2.21=pyhd8ed1ab_0
pydeprecate=0.3.2=pyhd8ed1ab_0
pyjwt=2.4.0=py38h06a4308_0
pyopenssl=22.1.0=pyhd8ed1ab_0
pyparsing=3.0.9=py38h06a4308_0
pyqt=5.15.7=py38h6a678d5_1
pyqt5-sip=12.11.0=py38h6a678d5_1
pysocks=1.7.1=pyha2e5f31_6
pytables=3.8.0=py38hb8ae3fc_3
python=3.8.13=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-lmdb=1.4.1=py38h6a678d5_0
python_abi=3.8=2_cp38
pytorch=1.12.1=py3.8_cuda11.6_cudnn8.3.2_0
pytorch-lightning=1.7.7=pyhd8ed1ab_0
pytorch-mutex=1.0=cuda
pytz=2022.1=py38h06a4308_0
pywavelets=1.3.0=py38h7f8727e_0
pyyaml=6.0=py38h7f8727e_1
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
readline=8.1.2=h7f8727e_1
reproc=14.2.3=h7f98852_0
reproc-cpp=14.2.3=h9c3ff4c_0
requests=2.28.1=pyhd8ed1ab_1
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
ruamel_yaml=0.15.80=py38h0a891b7_1007
scikit-image=0.19.2=py38h51133e4_0
scikit-learn=1.1.2=py38h6a678d5_0
scipy=1.9.1=py38h14f4228_0
setuptools=58.0.4=py38h06a4308_0
sip=6.6.2=py38h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h295c915_0
sortedcontainers=2.4.0=pyhd3eb1b0_0
sqlite=3.39.3=h5082296_0
tblib=1.7.0=pyhd3eb1b0_0
tensorboard=2.10.1=pyhd8ed1ab_0
tensorboard-data-server=0.6.0=py38hca6d32c_0
tensorboard-plugin-wit=1.8.1=py38h06a4308_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2021.7.2=pyhd3eb1b0_2
timm=0.6.13=pyhd8ed1ab_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.11.2=pyhd3eb1b0_0
torchaudio=0.12.1=py38_cu116
torchmetrics=0.10.0=pyhd8ed1ab_0
torchvision=0.13.1=py38_cu116
tornado=6.2=py38h5eee18b_0
tqdm=4.64.1=pyhd8ed1ab_0
typing-extensions=4.4.0=hd8ed1ab_0
typing_extensions=4.4.0=pyha770c72_0
urllib3=1.26.11=pyhd8ed1ab_0
wcwidth=0.2.5=pyhd3eb1b0_0
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.6=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yaml-cpp=0.7.0=h27087fc_2
yarl=1.8.1=py38h5eee18b_0
zfp=0.5.5=h295c915_6
zict=3.0.0=py38h06a4308_0
zipp=3.8.0=py38h06a4308_0
zlib=1.2.12=h5eee18b_3
zlib-ng=2.0.7=h5eee18b_0
zstd=1.5.2=ha4553b6_0
